first commit from EC2

+++ app.py
@@ -3,24 +3,25 @@ from flask_sqlalchemy import SQLAlchemy
 
 app = Flask(__name__)
 
-# /// = relative path, //// = absolute path
+# Database config
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///db.sqlite'
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+
 db = SQLAlchemy(app)
 
-
+# Model
 class Todo(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     title = db.Column(db.String(100))
-    complete = db.Column(db.Boolean)
+    complete = db.Column(db.Boolean, default=False)
 
-
+# Home route
 @app.route("/")
 def home():
     todo_list = Todo.query.all()
     return render_template("base.html", todo_list=todo_list)
 
-
+# Add todo
 @app.route("/add", methods=["POST"])
 def add():
     title = request.form.get("title")
@@ -29,22 +30,27 @@ def add():
     db.session.commit()
     return redirect(url_for("home"))
 
-
+# Update todo
 @app.route("/update/<int:todo_id>")
 def update(todo_id):
-    todo = Todo.query.filter_by(id=todo_id).first()
-    todo.complete = not todo.complete
-    db.session.commit()
+    todo = Todo.query.get(todo_id)
+    if todo:
+        todo.complete = not todo.complete
+        db.session.commit()
     return redirect(url_for("home"))
 
-
+# Delete todo
 @app.route("/delete/<int:todo_id>")
 def delete(todo_id):
-    todo = Todo.query.filter_by(id=todo_id).first()
-    db.session.delete(todo)
-    db.session.commit()
+    todo = Todo.query.get(todo_id)
+    if todo:
+        db.session.delete(todo)
+        db.session.commit()
     return redirect(url_for("home"))
 
+# Run app
 if __name__ == "__main__":
-    db.create_all()
-    app.run(debug=True)
+    with app.app_context():
+        db.create_all()
+
+    app.run(host="0.0.0.0", port=5000, debug=True)
